[#231] Refactor: Clean up all index.ts files from Metrics
Cleaned up index.ts files from the following folders:
  - /Metrics/components
  - /Metrics/hooks
  - /Metrics/utils
