--- src/brokers/broker-details/components/Overview/Metrics/components/CardBodyLoading/index.ts
@@ -1,1 +1,0 @@
-export * from './CardBodyLoading';
