+++ src/brokers/broker-details/components/Overview/Metrics/components/CardBrokerCPUUsageMetrics/CardBrokerCPUUsageMetrics.container.tsx
@@ -1,11 +1,9 @@
 import { FC, useEffect, useState } from 'react';
-//import { CardBrokerCPUUsageMetrics } from './CardBrokerCPUUsageMetrics';
-import { parsePrometheusDuration } from '../../../../Overview/Metrics/utils';
-import { getMaxSamplesForSpan, valueFormatter } from '../../utils';
-import { useFetchCpuUsageMetrics } from '../../hooks';
+import { parsePrometheusDuration } from '../../../../Overview/Metrics/utils/prometheus';
+import { getMaxSamplesForSpan, valueFormatter } from '../../utils/format';
+import { useFetchCpuUsageMetrics } from '../../hooks/useFetchCpuUsageMetrics';
 import { useTranslation } from '../../../../../../../i18n/i18n';
 import { CardQueryBrowser } from '../CardQueryBrowser/CardQueryBrowser';
-//import { MetricsType } from '../../utils';
 
 export type CardBrokerCPUUsageMetricsContainerProps = {
   name: string;
