--- src/brokers/broker-details/components/Overview/Metrics/components/CardBrokerCPUUsageMetrics/index.ts
@@ -1,2 +1,0 @@
-//export * from './CardBrokerCPUUsageMetrics';
-export * from './CardBrokerCPUUsageMetrics.container';
