+++ src/brokers/broker-details/components/Overview/Metrics/components/CardBrokerMemoryUsageMetrics/CardBrokerMemoryUsageMetrics.tsx
@@ -7,7 +7,7 @@ import {
 import { useTranslation } from '../../../../../../../i18n/i18n';
 import { CardBodyLoading } from '../CardBodyLoading/CardBodyLoading';
 import { EmptyStateNoMetricsData } from '../EmptyStateNoMetricsData/EmptyStateNoMetricsData';
-import { ChartTitle } from '../ChartTitle';
+import { ChartTitle } from '../ChartTitle/ChartTitle';
 
 export type CardBrokerMemoryUsageMetricsProps = ChartMemoryUsageProps & {
   isInitialLoading: boolean;
