--- src/brokers/broker-details/components/Overview/Metrics/components/CardBrokerMemoryUsageMetrics/index.ts
@@ -1,2 +1,0 @@
-//export * from './CardBrokerMemoryUsageMetrics';
-export * from './CardBrokerMemoryUsageMetrics.container';
