--- src/brokers/broker-details/components/Overview/Metrics/components/ChartCPUUsage/index.ts
@@ -1,1 +1,0 @@
-//export * from './ChartCPUUsage';
