--- src/brokers/broker-details/components/Overview/Metrics/components/ChartMemoryUsage/index.ts
@@ -1,1 +1,0 @@
-//export * from './ChartMemoryUsage';
