--- src/brokers/broker-details/components/Overview/Metrics/components/ChartPopover/index.ts
@@ -1,1 +1,0 @@
-export * from './ChartPopover';
