--- src/brokers/broker-details/components/Overview/Metrics/components/ChartSkeletonLoader/index.ts
@@ -1,1 +1,0 @@
-export * from './ChartSkeletonLoader';
