--- src/brokers/broker-details/components/Overview/Metrics/components/ChartTitle/index.tsx
@@ -1,1 +1,0 @@
-export * from './ChartTitle';
