--- src/brokers/broker-details/components/Overview/Metrics/components/DropDownWithToggle/index.ts
@@ -1,1 +1,0 @@
-export * from './DropdownWithToggle';
