--- src/brokers/broker-details/components/Overview/Metrics/components/EmptyStateNoMetricsData/index.ts
@@ -1,1 +1,0 @@
-export * from './EmptyStateNoMetricsData';
