+++ src/brokers/broker-details/components/Overview/Metrics/components/MetricsActions/MetricsActions.tsx
@@ -1,8 +1,11 @@
 import { FC } from 'react';
 import { Card, CardActions, CardHeader } from '@patternfly/react-core';
 import { useTranslation } from '../../../../../../../i18n/i18n';
-import { MetricsType } from '../../utils';
-import { DropdownWithToggle, IDropdownOption } from '../DropDownWithToggle';
+import { MetricsType } from '../../utils/types';
+import {
+  DropdownWithToggle,
+  IDropdownOption,
+} from '../DropDownWithToggle/DropdownWithToggle';
 
 export type MetricsActionsProps = {
   pollingTime: string;
