--- src/brokers/broker-details/components/Overview/Metrics/components/MetricsActions/index.ts
@@ -1,1 +1,0 @@
-export * from './MetricsActions';
