--- src/brokers/broker-details/components/Overview/Metrics/components/MetricsLayout/index.ts
@@ -1,1 +1,0 @@
-export * from './MetricsLayout';
