--- src/brokers/broker-details/components/Overview/Metrics/components/index.ts
@@ -1,4 +1,0 @@
-export * from './CardBodyLoading';
-export * from './CardBrokerMemoryUsageMetrics';
-export * from './CardBrokerCPUUsageMetrics';
-export * from './MetricsActions';
