--- src/brokers/broker-details/components/Overview/Metrics/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './useChartWidth';
-export * from './useFetchMemoryUsageMetrics';
-export * from './useFetchCpuUsageMetrics';
