+++ src/brokers/broker-details/components/Overview/Metrics/hooks/useFetchMemoryUsageMetrics.tsx
@@ -3,7 +3,7 @@ import {
   usePrometheusPoll,
   PrometheusEndpoint,
 } from '@openshift-console/dynamic-plugin-sdk';
-import { QueryInput } from '../utils';
+import { QueryInput } from '../utils/types';
 
 export const memoryUsageQuery = (
   name: string,
