--- src/brokers/broker-details/components/Overview/Metrics/index.ts
@@ -1,1 +1,0 @@
-export * from './Metrics';
