--- src/brokers/broker-details/components/Overview/Metrics/utils/index.ts
@@ -1,8 +1,0 @@
-export * from './chart-theme';
-export * from './prometheus';
-export * from './format';
-export * from './units';
-export * from './types';
-export * from './data-utils';
-export * from './consts';
-export * from './datatime';
